lddfiles: print dependencies of executables
We've adopted a principle for u-root: if a command has the same name
as a gnubin command it should act like that command.

Given that ld.so does what we want, and ldd as defined by gnu is
not that useful in many ways, we're moving ldd to lddfiles and making it useful.

lddfiles will print out the executables and all dependencies, suitable for
use in scripts:
scp `lddfiles /bin/date` root@host:/
or for cppio
lddfiles `which xterm` | cpio -H newc -o > /tmp/x
